Move activity detection to its own file
