+++ src/main/index.ts
@@ -5,6 +5,7 @@ import { Preferences, Theme, preferencesStore } from '../common/preferences';
 import fs from "fs"
 import { protocol } from "electron";
 import { uIOhook } from 'uiohook-napi'
+import { ActivityDetection } from './activityDetector';
 
 let tray: any = null;
 let win: any = null;
@@ -79,7 +80,7 @@ const createWindow = () => {
         win.hide()
     });
 
-    activityDetector = new ActivityDetection();
+    activityDetector = new ActivityDetection(win);
 
     createModal();
 }
@@ -142,57 +143,6 @@ function loadHtml(window: any, fileName: string) {
     window.loadURL(`${process.env['ELECTRON_RENDERER_URL']}/${fileName}.html`)
   else
     window.loadFile(join(__dirname, `../renderer/${fileName}.html`))
-}
-
-const interactionIntervalMs = 4 * 1000;
-const numInteractionIntervals = 10;
-const minInteractionIntervals = 6;
-class ActivityDetection {
-  // if input has occured within at least x of y intervals that are z seconds long,
-  // then the user is considered active
-  interactionDateTime: Date | undefined = undefined;
-  hitInteractionTotal: number = 1;
-  interactionTotal: number = 1;
-
-  constructor() {
-    uIOhook.on("input", (_) => {
-      if(!this.interactionDateTime) {
-        this.interactionDateTime = new Date(new Date().valueOf() + interactionIntervalMs);
-        return;
-      }
-
-      if(win.isFocused()) {
-        this.interactionDateTime = new Date(new Date().valueOf() + interactionIntervalMs);
-        this.hitInteractionTotal = 1;
-        this.interactionTotal = 1;
-        return;
-      }
-  
-      const timeDelta = new Date().valueOf() - this.interactionDateTime.valueOf();
-      if(timeDelta >= 0) {
-        ++this.hitInteractionTotal;
-        this.interactionTotal += Math.floor(timeDelta / interactionIntervalMs) + 1;
-        if(this.hitInteractionTotal >= minInteractionIntervals) {
-          win.webContents.send("continuous-activity");
-          this.hitInteractionTotal = 1;
-          this.interactionTotal = 1;
-        } else if(this.interactionTotal >= numInteractionIntervals) {
-          this.hitInteractionTotal = 1;
-          this.interactionTotal = 1;
-        }
-        this.interactionDateTime = new Date(new Date().valueOf() + interactionIntervalMs);
-      }
-    });
-
-    if(preferencesStore.get("activityTracking"))
-      uIOhook.start();
-  }
-
-  reset() {
-    this.interactionDateTime = new Date();
-    this.hitInteractionTotal = 1;
-    this.interactionTotal = 1;
-  }
 }
 
 function createModal() {
